+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Added `Hex::as_u64` and `Hex::from_u64`
+
 ## 0.17.0
 
 * Added `HexagonalMap` storage structure for dense, hexagon shaped maps (#163)